pyproject.toml
[project]
name = "azure-llm-toolkit"
version = "0.1.0"
description = "Azure OpenAI client wrapper with rate limiting, cost tracking, and retry logic"
readme = "README.md"
authors = [
    { name = "Torstein Sørnes", email = "t.soernes@gmail.com" }
]
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["azure", "openai", "llm", "rate-limiting", "cost-tracking", "embeddings", "chat"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "polars>=0.19.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.8.0",
    "basedpyright>=1.31.4",
]

[project.urls]
Homepage = "https://github.com/torsteinsornes/azure-llm-toolkit"
Repository = "https://github.com/torsteinsornes/azure-llm-toolkit"
Issues = "https://github.com/torsteinsornes/azure-llm-toolkit/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_llm_toolkit"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.basedpyright]
typeCheckingMode = "standard"
